pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "chrysalis"
version = "0.1.0"
description = "Interdisciplinary research framework studying criticality and transformation dynamics"
requires-python = ">=3.11"
license = "Apache-2.0"
dependencies = [
    "numpy>=1.24",
    "matplotlib>=3.7",
    "numba>=0.60",
    "networkx>=3.1",
    "pandas>=2.0",
    "PyYAML>=6.0",
    "jsonschema>=4.0",
    "typer>=0.9",
    "mcp[cli]",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "wordcloud>=1.8",
]

[project.scripts]
chrysalis-parse = "src.tools.supernote_parser:app"
chrysalis-validate = "src.tools.validate_notes:app"

[tool.setuptools.packages.find]
include = ["src*", "simulations*"]
